+++ Sonic/src/Sonic/Scene/ComponentView.h
@@ -64,9 +64,9 @@ namespace Sonic {
 				function(*it);
 		}
 
-		size_t Size()
+		int Size()
 		{
-			return m_Pool->m_Size;
+			return static_cast<int>(m_Pool->m_Size);
 		}
 
 		ComponentPool* m_Pool;
